[pf-579] Rename SmartRandom to smartRandom
The Capital letter is confusing. Since a previous refactoring,
this file does not expert a class anymore but uses regular functions
