+++ api/lib/domain/services/smart-random/smartRandom.js
