+++ api/tests/unit/domain/usecases/get-next-challenge-for-competence-evaluation_test.js
@@ -1,7 +1,7 @@
 const { expect, sinon, domainBuilder, catchErr } = require('../../../test-helper');
 const { UserNotAuthorizedToAccessEntity } = require('../../../../lib/domain/errors');
 const getNextChallengeForCompetenceEvaluation = require('../../../../lib/domain/usecases/get-next-challenge-for-competence-evaluation');
-const SmartRandom = require('../../../../lib/domain/services/smart-random/SmartRandom');
+const SmartRandom = require('../../../../lib/domain/services/smart-random/smartRandom');
 
 describe('Unit | Domain | Use Cases | get-next-challenge-for-competence-evaluation', () => {
